Index series, not chunks
We should index the series, not chunks; this will reduce the number of entries in the index by `replication factor * (chunk size / bucket size)`, or 3 * 6hrs / 24hrs - ie 12x. This will however mean we need another index from series to chunks, introducing 1 extra write an N extra reads per query.  Expectation is a reduction in query latency (and bigtable query usage, and memory cost) by 12x, and then an increase by 2x as we have to do a bunch of queries.

This change introduces the seriesStore, a new chunk store implementation that, combined with the v9 schema, indexes series not chunks.

I tried to adapt the original chunk store to support this style of indexing - easy on the write path, but the read path became even more of a rats nest.  So I factored out the common bits as best I could and made a new chunk store.

Signed-off-by: Tom Wilkie <tom.wilkie@gmail.com>

Tidy up some of the logging.

Signed-off-by: Tom Wilkie <tom.wilkie@gmail.com>