Apply jitter to chunk ages so the queue doesn't jump instantly
When there are a lot of timeseries that go on for days, all of them
will age out at around the same time, 12 hours after the ingester
started.

This change smears out that timing, using the series fingerprint as a
handy random number, so the effect isn't so pronounced.  At the cost
of slightly more chunks saved.

Signed-off-by: Bryan Boreham <bryan@weave.works>