Split the index and chunk store so we can mix and match.
This PR allows us for have the index in Bigtable, but chunks in GCS; or the index in Cassandra, and chunks in S3 etc.

Splitting the responsibilities has also reduce the amount of embedding in the various clients and made responsibilites clearer.

Also, unify the parallel fetching of multiple chunks between the Cassandra and S3 stores.

Signed-off-by: Tom Wilkie <tom.wilkie@gmail.com>