Allow ingesters to hand over chunks in example config
By setting `-ingester.claim-on-rollout=true`.

Note that the bit in this file about stopping one ingester before
starting another doesn't contradict this - Kubernetes will send
SIGTERM to the old one before starting the new one, but it doesn't
wait for it to stop. SIGTERM triggers the old ingester to look for a
new one to hand over to.

Signed-off-by: Bryan Boreham <bryan@weave.works>