Allow extra space to add new bigchunk data after Unmarshal
After a transfer of chunks from one ingester to another, a new
smallChunk is created immediately for any new samples that arrive.
Allowing room for this avoids a copy and doubling in size of the slice
of smallChunks.

Unfortunately we go through the same code path on queries, so they
will get slightly larger in memory - 24 bytes per chunk.

Signed-off-by: Bryan Boreham <bryan@weave.works>