Support consistent hashing in cache.NewMemcachedClient (#1554)
* chunk/cache: support consistent hashing in cache.NewMemcachedClient

cache.MemcachedClientConfig has been updated with a new boolean
variable ConsistentHash, available as consistent_hash in yaml and
memcached.consistent-hash as a flag.

When ConsistentHash is true, the MemcachedClient will use the
newly created cache.MemcachedJumpHashSelector for server distribution.

Jump hash is a consistent hashing algorithm that given a key and a
number of buckets, returns a bucket number in the range [0, numBuckets).
Adding or removing a bucket only results in 1/N keys being moved.

A downside to using jump hash is that buckets can not be arbitrarily
removed from the list; it effectively acts as a stack and only supports
adding or removing buckets from the end. Therefore, jump hash is most
effective when the servers are ordered and where the order is
predicable. A good example of this is Kubernete's StatefulSet with a
headless service. DNS names will be in the form memcached-[pod number],
where the pod number will grow and shrink in the way that numBuckets
does. There will never be a gap in the servers when scaling up or down.

Signed-off-by: Robert Fratto <robert.fratto@grafana.com>

* chunk/cache: Add natural sort in jump hash server selector

Signed-off-by: Robert Fratto <robert.fratto@grafana.com>

* chunk/cache: add test to demonstrate natsort works

TestNatSort has been added to validate that the natsort package works as
expected when sorting a list of servers that are returned from SRV
lookups. The example used corresponds to SRV records that would be
returned for a k8s headless service backed by a StatefulSet.

Signed-off-by: Robert Fratto <robert.fratto@grafana.com>

* Add changelog entry.

Signed-off-by: Tom Wilkie <tom.wilkie@gmail.com>