Simplify long-term caching
An attempt to fix https://github.com/cortexproject/cortex/issues/1698

We don't mix things when the time-range for the query overlaps the
"active" time-range. We consider all index entries as active. This is
because the `IndexQuery` fields depend on the `from` value and changing
it might mess things up.

This is kinda only effective when paired with query-frontend as most
queries issued fall in the active-range, but the query-frontend with
it's splitting would make sure the queriers actually only see some
queries that are totally in the non-active range.

Signed-off-by: Goutham Veeramachaneni <gouthamve@gmail.com>