Adds v11 schema (#1538)
* Adds v11 schema to store label names within index.

Stores only label names and not the entire metric. Storing entire metric
will bloat the index by 30% and it doesn't really make sense to do it
right now. Adding just label names adds a tolerable 7% to the index.

Also, in Prometheus, we don't treat __name__ as a special label. We
always return it when calling /labels API and we should do the same
here.

Signed-off-by: Cyril Tovena <cyril.tovena@gmail.com>
Signed-off-by: Goutham Veeramachaneni <gouthamve@gmail.com>

* Finishing touches

fix lint issue

Signed-off-by: Cyril Tovena <cyril.tovena@gmail.com>

removes useless loop

Signed-off-by: Cyril Tovena <cyril.tovena@gmail.com>

This should be on v11 not v10.

Signed-off-by: Cyril Tovena <cyril.tovena@gmail.com>

s/metricConstRangeKeyV1/labelNamesRangeKeyV1/

The code was first written to store the entire series, but now changed
to do just labelNames.

Signed-off-by: Goutham Veeramachaneni <gouthamve@gmail.com>

Add note about v11 being experimental.

Signed-off-by: Goutham Veeramachaneni <gouthamve@gmail.com>
Co-authored-by: cyriltovena <cyril.tovena@gmail.com>
Co-authored-by: Goutham Veeramachaneni <gouthamve@gmail.com>