Apply jitter to the configured timeout for performing idle compactions. (#3850)
* Apply jitter to the configured timeout for performing idle compactions.

Attempt to prevent idle compaction from happening concurrently by
introducing a 25% jitter to the configured timeout. This jitter only
increases the timeout, such that the configured minimum timeout is
always adhered to.

Signed-off-by: Steve Simpson <steve.simpson@grafana.com>

* Update Changelog.

Signed-off-by: Steve Simpson <steve.simpson@grafana.com>

* Fix race detected in unit test TestIngesterCompactIdleBlock.

Move the setting of TSDB.compactionIdleTimeout just after struct
creation. This avoids a race condition in the unit test because it makes
calls to compactBlocks directly.

Signed-off-by: Steve Simpson <steve.simpson@grafana.com>

* Update configuration flag documentation.

Signed-off-by: Steve Simpson <steve.simpson@grafana.com>

Co-authored-by: Marco Pracucci <marco@pracucci.com>
Former-commit-id: 5a2ba416ede1b2730b5f032cffcf8fcca1908053