Wait for alertmanagers to complete state sync before becoming ACTIVE. (#4161)
* Wait for alertmanagers to complete state sync before becoming ACTIVE.

Note: Applies to sharding operation only.

When starting up, make sure that the initial state sync is completed for
any users which were initially assigned to the instance. This will reduce
the possibility that we lose state when [cleanly] scaling up or down, by
inadvertently losing the replica which we need to obtain state from.

This change allows the state sync to progress concurrently for all
tenants. The alternative would be to block for the service in `New()`,
which is a slightly simpler code change, but would mean that start-up
is potentially much slower.

Signed-off-by: Steve Simpson <steve.simpson@grafana.com>

* Review comments.

Signed-off-by: Steve Simpson <steve.simpson@grafana.com>
Former-commit-id: 76b12a44e3505c7740e0382d5b4054c41a3710f2