Delete alertmanager state objects from remote storage on user deletion. (#4167)
* Delete alertmanager state objects from remote storage on user deletion.

The same pattern as for cleaning up local state is used - checking for
state objects to clean up whenever the users are synced, and deleting
them if the user has no configuration. This means that we perform an
additional listing operation as opposed to doing it on demand, when the
user configuration is deleted. This does mean that the deletion will be
more robust against e.g. the process dying before cleaning up, or
transient storage errors causing deletion to fail.

Note that a different mechanism is used to check if a remote user state
should be deleted, compared to for local state, because every user may
not be currently active in every alertmanager instance due to sharding.
Therefore, we only delete state for users which have also have no
configuration. The user listing is re-used from the user sync.

Signed-off-by: Steve Simpson <steve.simpson@grafana.com>

* Fix deletion to use correct user ID listing and add unit test.

Signed-off-by: Steve Simpson <steve.simpson@grafana.com>

* Review comments.

Signed-off-by: Steve Simpson <steve.simpson@grafana.com>
Former-commit-id: 324c265b53de03ca8f101d1d30258128d6c42958