Fix StateReplicationWithSharding_InitialSyncFromPeers flaking. (#4349)
The test is writing a single silence and  checking a metric which
indicates whether replicating the silence has been attempted yet.
This is so we can check later on that no replication activity
occurs. The assertions later on in the test are passing, but the
first one is not, indicating that the replication doesn't trigger
early enough. This makes sense because the replication is not
synchronous with the writing of the silence.

Signed-off-by: Steve Simpson <steve.simpson@grafana.com>
Former-commit-id: 65466d24acf4afb94506f28572be0c03a813e0af