Refactor NewMemcachedClient constructor (#483)
* Refactor NewMemcachedClient constructor

Previously NewMemcachedClient was returning the MemcachedClient
interface that only had a subset of cache methods. This was
limiting the use of the constructor to a single type of use.

The introduced change by this commit renames the constructor to
NewUpdatedMemcachedClient and refactors the return type that replaces
the interface with a specific type.

This enables the the constructor to be used for other purposes where the
caller may want to use other available methods in the memcached client
like Add, Delete, CompareAndSwap etcetera.

* Rename MemcacheClient to GetterSetterMemcached

The interface was called MemcachedClient which conflicts with the actual
type that wraps the memcached client. So the change of this commit is to
rename the MemcachedClient to GetterSetterMemcached which gives more
information about what the interface actually does.

In a follow up commit we'll rename the UpdateMemcachedClient type to
simply MemcachedClient.

* Renames UpdatedMemcachedClient to MemcachedClient

After the conversation in
https://github.com/grafana/mimir/pull/483#discussion_r747488755 we've
decided to rename the type UpdatedMemcachedClient to simply
MemcachedClient which now does not conflict with the name of the
GetterSetterMemcached interface.

* Stop embedding memcache client and actually implement methods

Before the refactor we did in the last two commits, we were only
exposing two memcache methods which were Set() and GetMulti(). With the
refactor and because MemcachedClient{} had a memcache client embedded we
were exposing any available methods in the client.

To make the behaviour consistent instead of embedding the memcache
client we're now defining the methods that we really want to expose.
Also this allows us to extend the behaviour like in the Set() method
where we're actually limiting the max item set or error wrapping where
needed.

* MemcachedClient doesnt embed sync.Mutex anymore

MemcachedClient was anonymously embedding sync.Mutex interface. The
change introduced in this commit stops this behaviour because we don't
want to have sync.Mutex methods as part of this structs public API.

* Implement new methods for MemcachedClient

In order to provide a more complete MemcachedClient I've implemented
Add(), CompareAndSwap(), Get() and Delete(). This will remain unused in
the exisitng codebase but are available to offer a more complete public
API for any possible caller.

Also all the methods above are consistently limiting the max item size
and error wrapping the responses in case of error.

* Update memcached client error wrapping and conditions

CompareAndSwap now checks for max item size and if its exceeded it
fails.

Also Add now returns an error instead of nil if the item is exceeded.

Also extracted the server error wrapping to its own function so it can
be reused throughout all the methods.

* Reorder errors import

* Mutex hat for the circuitbreaking map list

The MemcachedClient struct had a mutex defined at the top that was only
used for handling concurrent writes and reads to the cirbuit breaker
map. We've decided to place the mutex next to this map in the struct so
that its easier for a reader to spot the relationship between the two
variables.