Querier: labels requests always fetch from store and respect time ranges (#518)
* querier: always fetch labels from store

Remove description of the feature as experimental; make the option
`-querier.query-store-for-labels-enabled` always on.

* Querier: labels requests obey query-ingesters-within

If `-querier.query-ingesters-within` is set, make series, labels and
label-values requests to ingesters only within that time.

Ingesters may have data going back much further, depending on retention
setting, so it's more efficient to query just the window we won't get
from blocks storage.

Historically we only queried ingesters, so it was good to make the
request as long as possible, but now we always fetch labels from block
storage we don't need to.

* Use RegisterFlags with logger

* Review feedback

* Fix up calls to RegisterFlags

In other tools and places that use the config struct.

* Fix up test to match new behaviour.

* Ingesters: respect start/end time on labels&series requests

Previously ingester would return all current data if you asked for an
old range. Now it will return data in range.

With the new behaviour, if you don't set `EndTimestampMs` then it
is treated as zero and you will get nothing back. This is on the
internal gRPC API and it different to the Prometheus API which treats no
end time as "till the end of time".

One test needed fixing as it actually requested data from the wrong
time - the zero value `time.Time{}` is a specific point in the past.
Make it request data for all time instead.

Fix up other tests to match the new behaviour.

* Removed duplicated CHANGELOG entry

Signed-off-by: Marco Pracucci <marco@pracucci.com>

Co-authored-by: Marco Pracucci <marco@pracucci.com>