Krajo/blocks store graceful query errorhandling (#1003)
* querier: retry or ignore failed store gateways

Do not fail the whole query if a store gateway fails in it. Let
 queryWithConsistencyCheck retry of successed if all blocks are present.

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>

* querier_test: follow up change to retry or ignore failed store gateways

Simply changing the expected error Detail is not enough, since consistency
check is not done for recently uploaded blocks. In real life we do not
ask store gateways for recent blocks, rather we ask ingesters, see
querier.query-ingesters-within option.

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>