fix running helm vs jsonnet compare on outside PRs (#2533)
* Fix running helm vs jsonnet compare on external PRs

It is not possible to automatically comment on external (to our public
mimir repo) PRs. As per documentation:

https://github.com/marketplace/actions/create-or-update-comment :
Note: In public repositories this action does not work in pull_request
workflows when triggered by forks. Any attempt will be met with the error,
Resource not accessible by integration. This is due to token restrictions
 put in place by GitHub Actions.

As a fix, we ignore such errors from community PRs. The output will be in
the github action log only.

* Helm-jsonnet diff only comment if there's a problem

Was too verbose for some.

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>