Baremetal dashboards (#2657)
* Dashboards: move resources queries to config.libsonnet

This is the preliminary step to make room for supporting baremetal
deployment dashboards.

Signed-off-by: Wilfried Roset <wilfriedroset@users.noreply.github.com>

* Dashboards: use mapping for string formating

Queries and their arguments are in different files, hence using mapping
for string formating helps for the readability. One other benefit is
that it avoid to pass multiple time the same value or to pass exactly
the same number of attributes.

Signed-off-by: Wilfried Roset <wilfriedroset@users.noreply.github.com>

* Dashboards: move network metrics name to config.libsonnet

This is the preliminary step to make room for supporting baremetal
deployment dashboards.

Signed-off-by: Wilfried Roset <wilfriedroset@users.noreply.github.com>

* Dashboards: Introduce support for baremetal deployment

The following commit introduce the support of baremetal deployment for
*resources* dashboards. It's done by completing the hash
`resources_panel_series` and `resources_panel_queries`. The promql
queries are based on node_exporter.

To accommodate panels where the number of query differ we rely on two
intermediate functions `resourcesPanelLegend` and
`resourcesPanelQueries` who are responsible for returning the proper
number of queries/legends.

Signed-off-by: Wilfried Roset <wilfriedroset@users.noreply.github.com>

* Dashboards: align the use of containerName vs instanceName

As the dashboards supports both k8s and baremetal deployments,
let's use one single term to avoid confusion, hence instanceName.

Signed-off-by: Wilfried Roset <wilfriedroset@users.noreply.github.com>

* Dashboards: deployment_type default value is kubernetes

Let's use a more meaning name for the deployment type. The metrics
relates to k8s metrics, as such having the value as kubernetes instead
of container makes more sense.

Signed-off-by: Wilfried Roset <wilfriedroset@users.noreply.github.com>

Signed-off-by: Wilfried Roset <wilfriedroset@users.noreply.github.com>