Make blocks cleanup in compactor more fair (#2631)
This is a proposal change to show how I am thinking of making
blocks cleanup more fair. The problem is described in more detail in
<issue>. The changes in pool.go fit better in
grafana.com/dskit/concurrency, but I wanted to get feedback on the
approach before submitting a PR in dskit. Tests will also be added in
dskit.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>