streaming store-gateway: avoid copying postings when hashing them (#3779)
This PR solves two problems:

a bug I just discovered - when we previously constructed the byte slice of the postings we were incorrectly setting the byte position of the octet from the posting
it was
hashable[i+octet]   = byte(posting >> (octet * 8))
it should be
hashable[i*8+octet] = byte(posting >> (octet * 8))
avoids copying the postings into a byte slice altogether. Instead, it creates a new sliceHeader that it points to the data of the postings and adjusts the length and capacity.
It also changes the hashing functions from SHA1 (20B) + Blake2 (32B) to only Blake2 (32B). The memcached cache key is too small to fit both SHA1 and Blake2. In another change that will also remove the matchers hash from the key I will use Blake2 with 42B